src/main/java/org/example/Account.java
package org.example;

public class Account {
    private final String name;

    public Account(String name) {
        this.name = name;
    }

    public boolean checkNameToEmboss() {
        /*
             Этот метод должен проверять, что сохранённая через конструктор строка соответствует требованиям.
             Если строка удовлетворяет условиям, метод возвращает true, иначе — false.
         */
        if (name.length() >= 3 && name.length() < 20) {
            if (name.equals(name.trim())) {
                return name.contains(" ");
            }
        }
        return false;
    }

}
